# doc/API_Specification_OAS3.yaml
openapi: "3.0.2"
info:
  version: 1.0.0
  title: "Organz API"
  description: |
    An API used to access the Human Harvest Inc server.
    Versioning: Adding new endpoints, adding parameters or key value pairs are all minor changes, as they are still valid with a past client.
      Major changes include: Modifying the data types of any parameters or values, removing key value pairs, removing endpoints. These CANNOT be pushed without maintaining a legacy version and upgrading the system to a new version.
...
paths:
...
  /administrators:
    get:
      tags:
        - Administrator
      security:
        - AdministratorAccess: []
      summary: Returns all administrators from the system
      parameters:
        - $ref: '#/components/parameters/q'
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/count'
      responses:
        200:
          description: A list of administrators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/administratorDetails'

        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: '#/components/responses/401Unauthorized'
        403:
          $ref: '#/components/responses/403Forbidden'
        500:
          $ref: '#/components/responses/500InternalServerError'
    post:
      tags:
        - Administrator
      security:
        - AdministratorAccess: []
      summary: Create a new administrator
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/administratorCreateDetails'
        required: true
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/administratorDetails'
        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: '#/components/responses/401Unauthorized'
        403:
          $ref: '#/components/responses/403Forbidden'
        500:
          $ref: '#/components/responses/500InternalServerError'
  '/administrators/{username}':
    parameters:
    - in: path
      name: username
      description: Administrator ID
      required: true
      schema:
        type: integer
        minimum: 0
    get:
      tags:
        - Administrator
      security:
        - AdministratorAccess: []
      summary: Get detailed administrator information
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/administratorDetails'
        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: '#/components/responses/401Unauthorized'
        403:
          $ref: '#/components/responses/403Forbidden'
        404:
          $ref: '#/components/responses/404NotFound'
        500:
          $ref: '#/components/responses/500InternalServerError'
    patch:
      tags:
        - Administrator
      security:
        - AdministratorAccess: []
      summary: Change some information about an administrator
      description: Only those properties included in the request body will be changed.
      parameters:
        - $ref: '#/components/parameters/etag'

      requestBody:
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/administratorModifiableDetails'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/administratorDetails'
        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: '#/components/responses/401Unauthorized'
        403:
          $ref: '#/components/responses/403Forbidden'
        404:
          $ref: '#/components/responses/404NotFound'
        412:
          $ref: '#/components/responses/412PreconditionFailed'
        428:
          $ref: '#/components/responses/428PreconditionRequired'
        500:
          $ref: '#/components/responses/500InternalServerError'
    delete:
      tags:
        - Administrator
      security:
        - AdministratorAccess: []
      summary: Delete an administrator
      parameters:
        - $ref: '#/components/parameters/etag'

      responses:
        201:
          description: Deleted
        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: '#/components/responses/401Unauthorized'
        403:
          $ref: '#/components/responses/403Forbidden'
        404:
          $ref: '#/components/responses/404NotFound'
        412:
          $ref: '#/components/responses/412PreconditionFailed'
        428:
          $ref: '#/components/responses/428PreconditionRequired'
        500:
          $ref: '#/components/responses/500InternalServerError'
  '/commands':
    post:
      tags:
        - Administrator
      security:
        - AdministratorAccess: []
      summary: Execute command line commands on the backend database. Try executing "help" for information
      requestBody:
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/command'
      responses:
        200:
          description: The response from the command
        400:
          $ref: '#/components/responses/400BadRequest'
        401:
          $ref: '#/components/responses/401Unauthorized'
        403:
          $ref: '#/components/responses/403Forbidden'
        500:
          $ref: '#/components/responses/500InternalServerError'
...
components:
  parameters:
    q:
      name: q
      in: query
      description: "The search string"
      schema:
        type: string
    offset:
      name: offset
      in: query
      description: "The offset from the start of the result set, used for pagination"
      schema:
        type: integer
    count:
      name: count
      in: query
      description: "How many results to return, used for pagination"
      schema:
        type: integer
...
    etag:
      name: etag
      in: header
      description: A hashed value of the object used for optimistic concurrency control
      required: true
      schema:
        $ref: '#/components/schemas/etag'
...
  schemas:
...
    administratorDetails:
      type: object
      properties:
        username:
          type: string
        createdOn:
          $ref: '#/components/schemas/dateTime'
        modifiedOn:
          $ref: '#/components/schemas/dateTime'
    administratorCreateDetails:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    administratorModifiableDetails:
      type: object
      properties:
        username:
          type: string



    #Client attributes
...
    dateTime:
      type: string
      example: "2018-07-19T14:56:01.562"
...
    command:
      type: object
      properties:
        command:
          type: string
          example: help
    etag:
      type: string
      description: A hashed value of the object used for optimistic concurrency control
  responses:
    400BadRequest:
      description: Bad request
    401Unauthorized:
      description: Access token is missing or invalid
    403Forbidden:
      description: You do not have permission to perform that action
    404NotFound:
      description: The specified resource was not found
    412PreconditionFailed:
      description: That resource has already been modified or the ETag header is invalid
    428PreconditionRequired:
      description: ETag header was missing and is required to modify a resource
    500InternalServerError:
      description: Internal server error
  securitySchemes:
          
    ClientAccess:
      description: A client may only access their own data, they cannot access other users. They may only modify their personal information and organ donation status. They can view all their other statuses.
      type: http
      scheme: bearer
      bearerFormat: JWT
      
    ClinicianAccess:
      description: Clincians can access and modify and client, and may view any clinician, but only modify their own information.
      type: http
      scheme: bearer
      bearerFormat: JWT
      
    AdministratorAccess:
      description: Administrators can access and modify all user types.
      type: http
      scheme: bearer
      bearerFormat: JWT